implement category controller

+++ service/category_service.go
@@ -8,7 +8,7 @@ import (
 type CategoryService interface {
 	Create(ctx context.Context, request web.CategoryCreateRequest) web.CategoryResponse
 	Update(ctx context.Context, request web.CategoryUpdateRequest) web.CategoryResponse
-	Delete(ctx context.Context, categoryId int64)
-	FindById(ctx context.Context, categoryId int64) web.CategoryResponse
+	Delete(ctx context.Context, categoryId int)
+	FindById(ctx context.Context, categoryId int) web.CategoryResponse
 	FindAll(ctx context.Context) []web.CategoryResponse
 }
